Automatic Data Classifier

Fits multiple photosynthesis–irradiance (PI) models to the data in parallel and infers the data type based on the best adjusted R-squared.

DataType_piCurve(data, n_cores = 2)

Arguments

data

A data frame containing I (irradiance) and P (photosynthesis rate) columns.

n_cores

Integer. Number of CPU cores to use for parallel model fitting. Default is 2 to comply with CRAN check limits.

Value

A list with a single named element data_type, indicating the data type: ll (light-limited), ls (light-saturated), and ph (photoinhibition).

Details

The function fits a set of candidate models: photoinhibition models ("Ph09", "Ph10"), light-saturating models ("LS1", "LS5"), and a linear model for light-limited cases (models listed in Model_piCurve) and labels the data based on adjusted R2. R²adj is chosen over other statstical metric, as the PI data could be extremely imbalanced (see References for details).

If the linear model fits exceptionally well (R²adj > 0.94), it is strongly favored. For details on these model choices, see References.

References

Amirian M.M., V Finkel Z., Devred E., Irwin A.J., “A new parameterization of photoinhibition for phytoplankton, arXiv (2024) 1–33. 10.48550/arXiv.2412.17923.

Examples

# model parameters
params <- c(Pmax = 20, alpha = 0.6, beta = 0.1)

# generate an irradiance profile
df <- tibble::tibble(I = seq(0, 300, length = 25))

df$P <- # generate the photosynthetic rate using Jassby-tanh (LS5) model
 Model_piCurve(parameters = params, model_name = "2tanh", data = df) +
 2 * rnorm(25, 0, 0.25)

 DataType_piCurve(data = df)
#> $data_type
#> [1] "ph"
#>