Gregg and Carder (1990) atmospheric coefficients

Spectral atmospheric coefficients and extraterrestrial solar irradiance values used in the Gregg and Carder (1990) clear-sky maritime irradiance model.

GreggCarder1990

Format

A data frame with 351 rows and 5 variables:

lambda

Wavelength in vacuo (nm).

H_o

Spectral mean extraterrestrial solar irradiance (W m\(^{-2}\) nm\(^{-1}\)).

a_oz

Ozone absorption coefficient (cm\(^{-1}\)).

a_w

Water vapor absorption coefficient (cm\(^{-1}\)).

a_o

Oxygen absorption coefficient (cm\(^{-1}\)).

Source

Gregg, W. W., & Carder, K. L. (1990). A simple spectral solar irradiance model for cloudless maritime atmospheres. Journal of Geophysical Research.

Source values were transcribed into CSV format for use in piCurve.

Details

This dataset contains wavelength-specific coefficients used by the Gregg and Carder (1990) parameterization for estimating clear-sky spectral irradiance at the sea surface.

Examples

head(GreggCarder1990)
#>   lambda     H_o  a_oz a_w a_o
#> 1    350 0.92256 0.009 0.1 0.1
#> 2    351 0.91488 0.012 0.1 0.1
#> 3    352 0.91104 0.009 0.1 0.1
#> 4    353 1.01376 0.006 0.1 0.1
#> 5    354 1.07712 0.004 0.1 0.1
#> 6    355 1.03488 0.002 0.1 0.1