Estimate the information matrix for PI models

A wrapper to calculate the information matrix for PI models based on optimHess function.

InfoMat_piCurve(
  parameters,
  model_name,
  data,
  GradientFn = NULL,
  ...,
  Control = list()
)

Arguments

parameters

Vector -- containing the optimal values listed below:

Pmax \(\hspace{0.1cm}\): Maximum photosynthesis rate normalised by Chl_a (sign +),

alpha: Light-saturation slope at low light level (sign +),

beta \(\hspace{0.15cm}\): Photoinhibition rate at high light level (sign +),

R \(\hspace{0.7cm}\): dark reaction parameter (sign + OR -),

shape \(\hspace{0cm}\): shape parameter (sign +) -- Only required for some of the models. See Details.

StDev \(\hspace{0cm}\): standard deviation parameter -- Only required when STATapp = MLE.

model_name

String -- which model? (See details)

data

Vector -- Containing the irradiance profile.

GradientFn

String -- A function that produces the gradient for BFGS,

...

Further arguments to be passed to fn and gr.

Control

List -- a list of control parameters. Further details optim.

Value

A symmetric matrix that estimates the information/hessian matrix at the optimal solution.

Examples

# model parameters
params <- c(Pmax = 20, alpha = 0.6, R = 0, StDev = 2)

# generate an irradiance profile
df <- tibble::tibble(I = seq(0, 100, length = 25))

df$P <- # generate the photosynthetic rate using Jassby-tanh (LS5) model
   Model_piCurve(parameters = params, model_name = "tanh", data = df) +
   5 * rnorm(25, 0, 0.25)    # add some noise to P

# Estimate the optimal values for the generated dataset using MLE method
fit <- Fit_piModel(params, model_name = "tanh", STATapp = "MLE", data = df)

# Calculating the information matrix
InfoMat_piCurve(parameters = fit$par, model_name = fit$model, data = df)
#>               Pmax        alpha            R        StDev
#> Pmax   8.687425973 8.264574e+01 1.184187e+01  0.002149474
#> alpha 82.645740590 2.585497e+03 2.135084e+02  0.025067671
#> R     11.841867663 2.135084e+02 2.396700e+01  0.001686578
#> StDev  0.002149474 2.506767e-02 1.686578e-03 47.933660616

# different accuracy
InfoMat_piCurve(parameters = fit$par, model_name = fit$model, data = df,
                Control = list( ndeps = rep(1e-4, length = length(fit$par)) ) )
#>               Pmax        alpha            R        StDev
#> Pmax   8.687426600 8.264583e+01 1.184187e+01  0.002149037
#> alpha 82.645827248 2.585448e+03 2.135079e+02  0.018789059
#> R     11.841867753 2.135079e+02 2.396700e+01  0.001686473
#> StDev  0.002149037 1.878906e-02 1.686473e-03 47.933251324

InfoMat_piCurve(parameters = fit$par, model_name = fit$model, data = df,
               Control = list( ndeps = c(rep(1e-4, 3), StDev = 1e-3)) )
#>               Pmax        alpha            R        StDev
#> Pmax   8.687426600 8.264583e+01 1.184187e+01  0.002149214
#> alpha 82.645827248 2.585448e+03 2.135079e+02  0.018789237
#> R     11.841867753 2.135079e+02 2.396700e+01  0.001686562
#> StDev  0.002149214 1.878924e-02 1.686562e-03 47.933660616