Formulate PI Curve

Evaluate one of a large number of photosynthesis-irradiance (PI) models as a function of irradiance. Well-established functions such as bi-linear, exponential, hyperbolic tangent, second order polynomial, generalized rectangular hyperbola, and their appropriate combinations are provided.

Model_piCurve(parameters, model_name, data)

Arguments

parameters

Vector -- containing the values listed below:

Pmax \(\hspace{0.1cm}\): Maximum photosynthetic rate (sign +),

alpha: Slope of the light-saturation curve at low light (sign +),

beta \(\hspace{0.15cm}\): Photoinhibition parameter (sign +),

R \(\hspace{0.7cm}\): Dark reaction rate (sign + OR -),

shape \(\hspace{0cm}\): Shape parameter (sign +) -- Only required for some of the models. See Details.

model_name

String -- which model? (List of available models is given in details)

data

Vector -- containing the irradiance profile.

Value

A vector of predicted photosynthetic rate at each irradiances using the selected function.

Details

This package comprises three distinct model frameworks for formulating the PI curve: light-limited, light-saturating, and photoinhibition models. Each framework is detailed below

==============================

Table 1. Model used to formulate light-limited PI curve.

Name \(\hspace{0.5cm}\)	Model	Function Type	References

lm	\(P^B = \alpha I \hspace{1cm}\)	Linear	Linear Regression

==============================

Table 2. Models used to formulate light-saturating PI curve.

Name \(\hspace{0.5cm}\)	Model	Function Name	References

LS1	\(P^B = P^B_{\max}\dfrac{I+I_\alpha - |I-I_\alpha|}{2I_\alpha} \hspace{1cm}\)	Blackman-pw	Blackman 1905

LS2	\(P^B = P^B_{\max}\dfrac{I}{I + I_\alpha}\)	Baly-RH	Baly 1935

LS3	\(P^B = P^B_{\max}\dfrac{I}{\sqrt{I^2 + I^2_\alpha}}\)	Smith-RH	Smith 1936

LS4	\(P^B = P^B_{\max}(1-e^{-I/I_\alpha})\)	Webb-exp	Webb et al. 1974

LS5	\(P^B = P^B_{\max} \tanh \left( \dfrac{I}{I_\alpha} \right)\)	Jassby-tanh	Jassby et al. 1976

LS6	\(P^B = \dfrac{P^B_{\max}}{2 \theta} \left[\mathcal{I} -\sqrt{\mathcal{I}^2 - 4 \theta(\mathcal{I} - 1)} \right]\) \(\hspace{0.5cm}\)	Prioul-nonRH	Prioul et al. 1977

LS7	\(P^B = P^B_{\max} \dfrac{I}{(I^b+I^b_\alpha)^{1/b}}\)	Bannister-gRH \(\hspace{0.5cm}\)	Bannister 1979

==============================

Table 3. Models used to formulate photoinhibition PI curve.

Name \(\hspace{0.5cm}\)	Model	Function Name	References

Ph01	\(P^B = P^B_s~ (I/I_{\alpha}^s) ~e^{1-( I/ I_{\alpha}^s)} \hspace{1cm}\)	Steele-exp	Steele 1962

Ph02	\(P^B = P^B_s~ \dfrac{(I/I_{\alpha}^s)}{ ~ I^2 / (I_{\alpha}^s I_{\beta}^s) + (I/I_{\alpha}^s) + 1}\)	Peeters-rational	Peeters et al. 1978

Ph03	\(P^B = P^B_s~ (1-e^{-I/I_{\alpha}^s})e^{-I/I_{\beta}^s} \hspace{1cm}\)	Platt-exp	Platt et al. 1980

Ph04	\(P^B = P^B_s~ \tanh \left(I/I_{\alpha}^s\right) ~ e^{-I/I_{\beta}^s} \hspace{1cm}\)	Neale-exp	Neale et al. 1986

Ph05	\(P^B = P^B_s~ \dfrac{I}{I + I_{\alpha}^s} ~ e^{-I/I_{\beta}^s} \hspace{1cm}\)	Baly-exp	...

Ph06	\(P^B = P^B_s~ \dfrac{I}{\sqrt{I^2 + (I_{\alpha}^s)^2}} ~ e^{-I/I_{\beta}^s} \hspace{1cm}\)	Smith-exp	...

Ph07	\(P^B = P^B_s~ \dfrac{I}{(I^b + (I_{\alpha}^s)^b)^{1/b}} ~ e^{-I/I_{\beta}^s} \hspace{1cm}\)	Bannister-exp	...

Ph08	\(P^B = P^B_s~ \dfrac{P^B_{s}}{2 \theta} \left[ \mathcal{I} - \sqrt{\mathcal{I}^2 - 4 \theta(\mathcal{I} - 1)} \right] ~ e^{-I/I_{\beta}^s} \hspace{1cm}\)	Prioul-exp	...

Ph09	\(P^B = \begin{cases} \alpha I & \hspace{0.5cm} I \leq P^B_{\max} / \alpha \\ & \\ P^B_{\max} & \hspace{0.5cm} P^B_{\max} / \alpha < I \leq P^B_{\max} / \beta\\ & \\ -\beta I & \hspace{0.5cm} I > P^B_{\max} / \beta \end{cases} \hspace{1cm}\)	Extended-Blackman	...

Ph10	\(P^B = P^B_{\max} \tanh \left( \dfrac{I}{I_{\alpha}} \right) \tanh{ \left[ \left( \dfrac{I_{\beta}}{I} \right)^{\cosh^2(1)} \right] } \hspace{1cm}\)	Double-tanh	Amirian et al. 2024

Ph11	\(P^B = P^B_{\max} \tanh \left( \dfrac{I}{I_{\alpha}} \right) \tanh{ \left[ \left( \dfrac{I_{\beta}}{I} \right)^{\gamma} \right] } \hspace{1cm}\)	Double-tanh-shp	Amirian et al. 2024

Ph12	\(P^B = P^B_{\max} \left[1 - \exp{\left( -\dfrac{I}{I_{\alpha}} \right)} \right] \tanh{ \left[ \left( \dfrac{I_{\beta}}{I} \right)^{\cosh^2(1)} \right] } \hspace{1cm}\)	Exp-tanh	Amirian et al. 2024

Ph13	\(P^B = P^B_{\max} \left[1 - \exp{\left( -\dfrac{I}{I_{\alpha}} \right)} \right] \tanh{ \left[ \left( \dfrac{I_{\beta}}{I} \right)^{\gamma} \right] } \hspace{1cm}\)	Exp-tanh-shp	Amirian et al. 2024

Ph14	\(P^B = P^B_{\max} \tanh{\left(\dfrac{I}{I_{\alpha}}\right)} \left[1 - \exp{\left(-\dfrac{I_{\beta}}{I}\right)} \right] \hspace{1cm}\)	Tanh-rcp_exp	Amirian et al. 2024

Ph15	\(P^B = P^B_{\max} \left[1 - \exp{\left(-\dfrac{I}{I_{\alpha}}\right)} \right] \left[1 - \exp{\left(-\dfrac{I_{\beta}}{I}\right)} \right] \hspace{1cm}\)	Exp_rcp_exp	Amirian et al. 2024

Ph16	\(P^B = \dfrac{P^B_{\max}}{2 \theta} \left[1 + \theta_\beta ~ \mathcal{I}_\alpha -\sqrt{ \theta_\beta ~ \mathcal{I}_\alpha^2 - 4 \theta ~ \mathcal{I}_\alpha + 1} \right]\) \(\hspace{0.5cm}\)	Fasham-nonRH	Fasham et al. 1983

In equations LS6 and Ph08, \(b\) is a shape parameter, \(\theta = \left[ 1 + \exp{(-b)} \right]^{-1} \) is a sigmoid function setting \(0< \theta <1\), \(\theta_\beta = \theta + (1-\theta)\exp{(\beta I)}\), \(\mathcal{I} = \left({I}/{I_\alpha} + 1\right) \), and \(\mathcal{I}_\alpha = I / I_\alpha\). In the above tables also, \(I_{\alpha} = P_{\max}^B/ \alpha\), \(I^s_{\alpha} = P_{s}^B/ \alpha\), \(I_{\beta} = P_{\max}^B/ \beta\), and \(I^s_{\beta} = P_{s}^B/ \beta\).

==============================

Note that the default model_name for light-saturating and photoinhibition data types is Jassby-tanh (LS5) and double-tanh (Ph10) model, resepctively.

References

Amirian M.M., V Finkel Z., Devred E., Irwin A.J., “A new parameterization of photoinhibition for phytoplankton, arXiv (2024) 1–33. 10.48550/arXiv.2412.17923.

Examples

# model parameters
params <- c(Pmax = 20, alpha = 0.6, R = 0)

# generate an irradiance profile
df <- tibble::tibble(I = seq(0, 300, length = 25))

# compute the photosynthetic rate using Jassby-tanh (LS5) model
Model_piCurve(parameters = params, model_name = "tanh", data = df)
#>  [1]  0.000000  7.167148 12.702979 16.186021 18.102965 19.080905 19.560522
#>  [8] 19.791195 19.901095 19.953220 19.977889 19.989552 19.995064 19.997668
#> [15] 19.998899 19.999480 19.999754 19.999884 19.999945 19.999974 19.999988
#> [22] 19.999994 19.999997 19.999999 19.999999

# compute the photosynthetic rate using double-tanh (Ph10) model
Model_piCurve(parameters = c(params, beta = 0.3), model_name = "2tanh", data = df)
#>  [1]  0.0000000  7.1671480 12.7029790 16.1736674 17.4306718 15.7039817
#>  [7] 12.4871257  9.5053752  7.2322881  5.5868713  4.3988230  3.5287979
#> [13]  2.8794678  2.3853593  2.0024445  1.7007296  1.4594178  1.2638171
#> [19]  1.1033571  0.9703000  0.8588907  0.7647855  0.6846609  0.6159430
#> [25]  0.5566150