Find the optimal parameters in PI models

OptPIparams uses general-purpose optimization to identify the optimal parameters in the photosynthesis-irradiance curve models. The method includes the Nelder-Mead, quasi-Newton, and conjugate-gradient algorithms, as well as options for box-constrained optimization and simulated annealing. Two different statistical approaches, Mean Squared Error (MSE) and Maximum Likelihood Estimation (MLE), are available for objective function optimization.

OptPIparams(
  parameters,
  model_name = c("Eq1_lm", "Eq2_Blackman", "Eq3_Baly", "Eq4_Smith", "Eq5_Talling",
    "Eq6_Vollenweider_Ln", "Eq7_Webb", "Eq8_Jassby", "Eq9_Prioul", "Eq10_Bannister",
    "Eq11_Blackman_extended_1", "Eq12_Blackman_extended_2", "Eq13_Vollenweider",
    "Eq14_Steele", "Eq15_Steele_modified", "Eq16_Peeters", "Eq17_Platt", "Eq18_Neale",
    "Eq19_Neale_modified", "Eq20_Ye", "Eq21_Baly_extended", "Eq22_Smith_extended",
    "Eq23_Prioul_extended", "Eq24_Bannister_extended", "Eq25_Tanh_Tanh_1",
    "Eq26_Tanh_Tanh_2", "Eq27_Tanh_Tanh_shape", 
     "Eq28_Tanh_Expo",
    "Eq29_Expo_Tanh_1", "Eq30_Expo_Tanh_2", "Eq31_Expo_Tanh_shape", "Eq32_Expo_Expo",
    "Eq33_Amirian_Tanh", "Eq34_Amirian_Expo"),
  data,
  STATapp = c("MSE", "MLE"),
  GradientFn = NULL,
  ...,
  Method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  LowerBnd = -Inf,
  UpperBnd = Inf,
  Control = list(),
  Hessian = FALSE
)

Arguments

parameters

Vector -- containing the values listed below:

Pmax \(\hspace{0.1cm}\): Maximum photosynthesis rate normalised by Chl_a (sign +),

alpha: Light-saturation slope at low light level (sign +),

beta \(\hspace{0.15cm}\): Photoinhibition rate at high light level (sign +),

R \(\hspace{0.7cm}\): dark reaction parameter (sign + OR -),

shape \(\hspace{0cm}\): shape parameter (sign +) -- Only required for some of the models. See Details.

StDev \(\hspace{0cm}\): standard deviation parameter -- Only required when STATapp = MLE.

model_name

String -- which model? (See details)

data

Vector -- Containing the irradiance profile.

STATapp

String -- Providing the statistical methods for optimization, which are MSE or MLE. See Details. Default is MSE.

GradientFn

String -- A function that produces the gradient for BFGS, CG and L-BFGS-B techniques, and if it is not provided, a numerical estimation using finite-differences will be applied.

...

An option to pass further arguments to GradientFn.

Method

String -- The method employed to determine the optimal values for the model's parameters. Further details ?stats::optim.

LowerBnd

Numeric -- Lower boundaries on the variables for the L-BFGS-B method, or bounds in which to search for method Brent.

UpperBnd

Numeric -- Upper boundaries on the variables for the L-BFGS-B method, or bounds in which to search for method Brent.

Control

List -- a list of control parameters. Further details ?stats::optim.

Hessian

Logical -- Should the Hessian matrix be returned through numerical differentiation? Default is FALSE.

Value

Below is a list of the results obtained from the optimization.

par: The optimal parameters obtained through optimization by using either the mean squared error (MSE) or maximum likelihood estimate (MLE) method.

value: The MSE value on the primary production (PP) profile when STATapp = MSE is met or STATapp is not specified. This metric measures the average squared difference between predicted and actual values in regression analysis. A lower value for the mean MSE indicates a better fit. For STATapp = MLE, the optimization uses Maximum Likelihood Estimation method to find the optimal parameters. The value in this case stands for the likelihood of observing the primary production (PP) for the estimated parameters. A higher value for the MLE indicates a better fit as the method aims to find the parameter values that make the observed data most likely to occur according to the assumed distribution

counts: This is a vector consisting of two integers that indicate the number of times the model and its gradient GradientFn were called during a certain process. This count does not include the calls made to compute the Hessian (if requested), or any calls to the model to compute a finite-difference approximation to the gradient.

message: A character string that provides any extra information given by the optimizer upon completion of a process. if there is no such additional information, a value of NULL will be returned.

hessian: A symmetric matrix that estimates the Hessian at the optimal solution, but only if the hessian argument is set to TRUE. It is important to note that this is the Hessian of the unconstrained problem, even if there are active box constraints.

convergence: An integer value representing the optimization process is completed. A value of 0 indicates that the process has completed successfully, which is always the case for SANN and Brent. However, there may be other possible error codes listed below, which could be generated during the process.

1 : indicates that the iteration limit maxit had been reached.

10:indicates that the Nelder-Mead simplex method has encountered degeneracy, meaning that two or more vertices of the simplex have converged to the same point.

51:indicates a warning that has been issued by the L-BFGS-B method. Further details regarding this warning can be found in the message component of the output.

52: indicates an error that has occurred within the L-BFGS-B method. Additional details about this error can be found in the message component of the output.

info_criteria: A vector consisting AIC, AICc, and BIC values when MLE approach is chosen. The main difference between these values lie in the penalties they impose on model complexity. AIC tends to select more complex models compared to BIC, while AICc provides a correction to AIC for small sample sizes. BIC, on the other hand, penalizes model complexity more heavily than AIC and AICc, often resulting in the selection of simpler models.

model: Name of equation used to model the given data sample.

SQA: Statistical Quality Assessment consists of a vector incorporating R2 and adjusted R2 values. R2 values provide information about the goodness of the fit by indicating the proportion of data variation that is explained by the model. Adjusted R2 considers a penalty for over parameterization. Together, these two measures help evaluating the accuracy of the model's predictions and provide insights into how well the model captures/explains the observed data variation. Note that these values should not be used as a model selection criteria. AICc or BIC should be used for that purpose.

Details

The STATapp argument refers to the statistical approach. To achieve a more accurate estimation in a regression model (greater R\(^2\)), users may consider using the MSE method over the MLE method, as the former measures the average squared difference between predicted and actual values. In contrast, the MLE method determines the best possible values by maximizing the likelihood of observing the data within the given model, thus offering the user a detailed insight into the estimation error of each parameter. To get such insight when STATapp = MLE, the Hessian argument should be set to TRUE. If forgotten, the user can use optimHess function.

This package includes both commonly-used and recently-developed PI models. The list of models is provided below.

Table 1. Existing formulas for light-saturation curve (\(I_k = P_{\max}^B/ \alpha\)).

Name \(\hspace{0.5cm}\)	Model	Function Type	References

Eq1	\(P^B = P^B_{\max}\dfrac{I+I_k - |I-I_k|}{2I_k} \hspace{1cm}\)	Bilinear	Blackman 1905

Eq2	\(P^B = P^B_{\max}\dfrac{I}{I + I_k}\)	Rectangular Hyperbola	Baly 1935

Eq3	\(P^B = P^B_{\max}\dfrac{I}{\sqrt{I^2 + I^2_k}}\)	Modified Rectangular Hyperbola	Smith 1936

Eq4	\(P^B = P^B_{\max}(1-e^{-I/I_k})\)	Exponential	Webb et al. 1974

Eq5	\(P^B = P^B_{\max} \tanh \left( \dfrac{I}{I_k} \right)\)	Hyperbolic Tangent	Jassby et al. 1976

Eq6	\(P^B = \dfrac{P^B_{\max}}{2 \theta} \left[\mathcal{I} -\sqrt{\mathcal{I}^2 - 4 \theta(\mathcal{I} - 1)} \right]\) \(\hspace{0.5cm}\)	Non-rectangular hyperbola	Prioul et al. 1977

Eq7	\(P^B = P^B_{\max} \dfrac{I}{(I^b+I^b_k)^{1/b}}\)	Generalized Rectangular Hyperbola \(\hspace{0.5cm}\)	Bannister 1979

where \(b\) is a shape parameter, \(\theta = \dfrac{1}{1 + \exp{(-b)}}\) is a sigmoid function setting \(0< \theta <1\), and \(\mathcal{I} = \left(\dfrac{I}{I_k} + 1\right) \).

Note

The function Model_piCurve can handle all combinations of the names listed above, where the input string model_name is concerned. For instance,if any of the following forms are used, the function will set model_name = "Eq1_Blackman".

With or without spacing: Eq1Blackman, Eq1 Blackman,

Underscore or hyphen spacing: Eq1_Blackman, Eq1-Blackman,

Any form of lower/upper case: eq1_blackman, EQ1_blackman,

Missing author name: EQ1, Eq1, eQ1, eq1, 1,

Missing Eq indicator: blackman, Blackman,

Other combination of all the above: Blackmaneq1, BlackmanEq1, Blackman Eq1, eq1 Blackman, etc.

This function utilizes the optim() function from the stats package to perform the optimization. As such, the function makes available all of the arguments that can be used with the optim() function within itself.

Examples

# model parameters
params <- c(Pmax = 20, alpha = 0.6, R = 0)

# generate an irradiance profile
df <- tibble::tibble(I = seq(0, 100, length = 25))

df$PP <- # generate PP using Baly's rectangular hyperbola model
   Model_piCurve(parameters = params, model_name = "Eq3-Baly", data = df) +
   5 * rnorm(25, 0, 0.25)    # add some noise to PP

# Estimate the optimal values for the generated dataset using MSE method
OptPIparams(parameters = params, model_name = "Eq3-Baly", data = df)
#> $par
#>       Pmax      alpha          R 
#> 17.7914299  0.6328522  0.5638846 
#> 
#> $value
#> [1] 1.635195
#> 
#> $counts
#> function gradient 
#>      135       NA 
#> 
#> $convergence
#> [1] 0
#> 
#> $message
#> NULL
#> 
#> $model
#> [1] "Eq3-Baly"
#> 
#> $SQA
#>        R2     R2adj 
#> 0.8950952 0.8855584 
#> 

# Estimate the optimal values for the generated dataset using MLE method
OptPIparams(parameters = c(params, StDev = 2), model_name = "Eq3-Baly",
            STATapp = "MLE", Hessian = TRUE, data = df)
#> $par
#>       Pmax      alpha          R      StDev 
#> 17.7912095  0.6329431  0.5637845  1.2787091 
#> 
#> $value
#> [1] 41.62049
#> 
#> $counts
#> function gradient 
#>      215       NA 
#> 
#> $convergence
#> [1] 0
#> 
#> $message
#> NULL
#> 
#> $hessian
#>               Pmax        alpha            R        StDev
#> Pmax   2.580919567  32.80942610  5.569400889 -0.002547202
#> alpha 32.809426095 516.74022846 86.634834524 -0.036070269
#> R      5.569400889  86.63483452 15.289614034 -0.006026641
#> StDev -0.002547202  -0.03607027 -0.006026641 30.582141484
#> 
#> $info_criteria
#>      AIC     AICc      BIC 
#> 91.24097 93.24097 96.11648 
#> 
#> $model
#> [1] "Eq3-Baly"
#> 
#> $SQA
#>        R2     R2adj 
#> 0.8950952 0.8801088 
#>