Find the optimal parameters in PI models

OptPar_piCurve uses general-purpose optimization to identify the optimal parameters in the photosynthesis-irradiance curve models. The method includes the Nelder-Mead, quasi-Newton, and conjugate-gradient algorithms, as well as options for box-constrained optimization and simulated annealing. Two different statistical approaches, Mean Squared Error (MSE) and Maximum Likelihood Estimation (MLE), are available for objective function optimization.

OptPar_piCurve(
  parameters,
  model_name,
  data,
  data_type = c("light-limited", "light-saturating", "photoinhibition"),
  STATapp = c("MSE", "MLE"),
  GradientFn = NULL,
  ...,
  Method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  LowerBnd = -Inf,
  UpperBnd = Inf,
  Control = list(),
  Hessian = FALSE
)

Arguments

parameters

Vector -- containing the values listed below:

Pmax \(\hspace{0.1cm}\): Maximum photosynthetic rate normalised by Chl_a (sign +),

alpha: Light-saturation slope at low light level (sign +),

beta \(\hspace{0.15cm}\): Photoinhibition parameter (sign +),

R \(\hspace{0.7cm}\): Dark reaction rate (sign + OR -),

shape \(\hspace{0cm}\): Shape parameter (sign +) -- Only required for some of the models. See Details.

StDev \(\hspace{0cm}\): Standard deviation parameter -- Only required when STATapp = MLE.

model_name

String -- which model? (See details)

data

Vector -- Containing the irradiance profile.

data_type

String -- data sample type?

STATapp

String -- Providing the statistical methods for optimization, which are MSE or MLE. See Details. Default is MSE.

GradientFn

String -- A function that produces the gradient for BFGS, CG and L-BFGS-B techniques, and if it is not provided, a numerical estimation using finite-differences will be applied.

...

An option to pass further arguments to GradientFn.

Method

String -- The method employed to determine the optimal values for the model's parameters. Further details ?stats::optim.

LowerBnd

Numeric -- Lower boundaries on the variables for the L-BFGS-B method, or bounds in which to search for method Brent.

UpperBnd

Numeric -- Upper boundaries on the variables for the L-BFGS-B method, or bounds in which to search for method Brent.

Control

List -- a list of control parameters. Further details ?stats::optim.

Hessian

Logical -- Should the Hessian matrix be returned through numerical differentiation? Default is FALSE.

Value

Below is a list of the results obtained from the optimization.

par: The optimal parameters obtained through optimization by using either the mean squared error (MSE) or maximum likelihood estimate (MLE) method.

value: The MSE value on the primary production (PP) profile when STATapp = MSE is met or STATapp is not specified. This metric measures the average squared difference between predicted and actual values in regression analysis. A lower value for the mean MSE indicates a better fit. For STATapp = MLE, the optimization uses Maximum Likelihood Estimation method to find the optimal parameters. The value in this case stands for the likelihood of observing the primary production (PP) for the estimated parameters. A higher value for the MLE indicates a better fit as the method aims to find the parameter values that make the observed data most likely to occur according to the assumed distribution

counts: This is a vector consisting of two integers that indicate the number of times the model and its gradient GradientFn were called during a certain process. This count does not include the calls made to compute the Hessian (if requested), or any calls to the model to compute a finite-difference approximation to the gradient.

message: A character string that provides any extra information given by the optimizer upon completion of a process. if there is no such additional information, a value of NULL will be returned.

hessian: A symmetric matrix that estimates the Hessian at the optimal solution, but only if the hessian argument is set to TRUE. It is important to note that this is the Hessian of the unconstrained problem, even if there are active box constraints.

convergence: An integer value representing the optimization process is completed. A value of 0 indicates that the process has completed successfully, which is always the case for SANN and Brent. However, there may be other possible error codes listed below, which could be generated during the process.

1 : indicates that the iteration limit maxit had been reached.

10:indicates that the Nelder-Mead simplex method has encountered degeneracy, meaning that two or more vertices of the simplex have converged to the same point.

51:indicates a warning that has been issued by the L-BFGS-B method. Further details regarding this warning can be found in the message component of the output.

52: indicates an error that has occurred within the L-BFGS-B method. Additional details about this error can be found in the message component of the output.

info_criteria: A vector consisting AIC, AICc, and BIC values when MLE approach is chosen. The main difference between these values lie in the penalties they impose on model complexity. AIC tends to select more complex models compared to BIC, while AICc provides a correction to AIC for small sample sizes. BIC, on the other hand, penalizes model complexity more heavily than AIC and AICc, often resulting in the selection of simpler models.

model: Name of equation used to model the given data sample.

SQA: Statistical Quality Assessment consists of a vector incorporating R2 and adjusted R2 values. R2 values provide information about the goodness of the fit by indicating the proportion of data variation that is explained by the model. Adjusted R2 considers a penalty for over parameterization. Together, these two measures help evaluating the accuracy of the model's predictions and provide insights into how well the model captures/explains the observed data variation. Note that these values should not be used as a model selection criteria. AICc or BIC should be used for that purpose.

Details

The STATapp argument refers to the statistical approach. To achieve a more accurate estimation in a regression model (greater R\(^2\)), users may consider using the MSE method over the MLE method, as the former measures the average squared difference between predicted and actual values. In contrast, the MLE method determines the best possible values by maximizing the likelihood of observing the data within the given model, thus offering the user a detailed insight into the estimation error of each parameter. To get such insight when STATapp = MLE, the Hessian argument should be set to TRUE. If forgotten, the user can use optimHess function.

This package includes 36 models which can be found in ?piCurve::Model_piCurve both commonly-used and recently-developed PI models. The list of models is provided below.

This function utilizes the optim() function from the stats package to perform the optimization. As such, the function makes available all of the arguments that can be used with the optim() function within itself.

Examples

# model parameters
params <- c(Pmax = 20, alpha = 0.6, R = 0)

# generate an irradiance profile
df <- tibble::tibble(I = seq(0, 100, length = 25))

df$PP <- # generate the photosynthetic rate using Jassby-tanh (LS5) model
   Model_piCurve(parameters = params, model_name = "tanh", data = df, data_type = "ls") +
   5 * rnorm(25, 0, 0.25)    # add noise

# Estimate the optimal values for the generated dataset using MSE method
OptPar_piCurve(parameters = params, model_name = "tanh", data = df, data_type = "ls")
#> $par
#>       Pmax      alpha          R 
#> 18.1997025  0.5706693  0.9234946 
#> 
#> $value
#> [1] 1.605617
#> 
#> $counts
#> function gradient 
#>      143       NA 
#> 
#> $convergence
#> [1] 0
#> 
#> $message
#> NULL
#> 
#> $model
#> [1] "tanh"
#> 
#> $SQA
#>        R2     R2adj 
#> 0.9461993 0.9413083 
#> 

# Estimate the optimal values for the generated dataset using MLE method
OptPar_piCurve(parameters = c(params, StDev = 2), model_name = "tanh",
            STATapp = "MLE", Hessian = TRUE, data = df, data_type = "ls")
#> $par
#>       Pmax      alpha          R      StDev 
#> 18.2006727  0.5707439  0.9225903  1.2672191 
#> 
#> $value
#> [1] 41.39231
#> 
#> $counts
#> function gradient 
#>      231       NA 
#> 
#> $convergence
#> [1] 0
#> 
#> $message
#> NULL
#> 
#> $hessian
#>               Pmax         alpha             R        StDev
#> Pmax   6.860387778   42.13533262   8.715844052 -0.003288839
#> alpha 42.135332623 1020.53540644 103.273275144 -0.046162748
#> R      8.715844052  103.27327514  15.568134172 -0.005455373
#> StDev -0.003288839   -0.04616275  -0.005455373 31.129820361
#> 
#> $info_criteria
#>      AIC     AICc      BIC 
#> 90.78462 92.78462 95.66013 
#> 
#> $model
#> [1] "tanh"
#> 
#> $SQA
#>        R2     R2adj 
#> 0.9461993 0.9385135 
#>