Calculate astronomical day length

Computes day length (photoperiod) for a given date and latitude using solar declination calculated by solar_declination_piCurve().

day_length_piCurve(date, lat_deg, hour = 12)

Arguments

date

A Date, POSIXct, or POSIXt object.

lat_deg

Numeric latitude in degrees. Northern latitudes are positive and southern latitudes are negative.

hour

Numeric hour of the day used to calculate solar declination. Default is 12, representing local noon. Can be fractional.

Value

Numeric vector of day length in hours.

Details

Day length is calculated from the sunset hour angle:

$$ D = \frac{24}{\pi}\cos^{-1}(-\tan(\phi)\tan(\delta)) $$

where \(\phi\) is latitude and \(\delta\) is solar declination, both in radians. Polar day and polar night cases are handled by constraining the argument of \(\cos^{-1}\) to the interval \([-1, 1]\).

The hour argument is retained for consistency with solar_declination_piCurve(). For day-length calculations, hour = 12 is usually appropriate because within-day changes in solar declination are very small.

Examples

# Day length at Bedford Basin, Nova Scotia, on summer solstice
day_length_piCurve(as.Date("2026-06-21"), lat_deg = 44.69)
#> [1] 15.38552

# Seasonal day length over one year
dates <- seq(as.Date("2026-01-01"), as.Date("2026-12-31"), by = "day")
dl <- day_length_piCurve(dates, lat_deg = 44.69)