Formulate PI Curve

Wrapper to formulate the photosynthesis-irradiance curve using well-established functions such as bi-linear, exponential, hyperbolic tangent, second order polynomial, generalized rectangular hyperbola, and their appropriate combinations.

Model_piCurve(
  parameters,
  model_name = c("Eq1_lm", "Eq2_Blackman", "Eq3_Baly", "Eq4_Smith", "Eq5_Talling",
    "Eq6_Vollenweider_Ln", "Eq7_Webb", "Eq8_Jassby", "Eq9_Prioul", "Eq10_Bannister",
    "Eq11_Blackman_extended_1", "Eq12_Blackman_extended_2", "Eq13_Vollenweider",
    "Eq14_Steele", "Eq15_Steele_modified", "Eq16_Peeters", "Eq17_Platt", "Eq18_Neale",
    "Eq19_Neale_modified", "Eq20_Ye", "Eq21_Baly_extended", "Eq22_Smith_extended",
    "Eq23_Prioul_extended", "Eq24_Bannister_extended", "Eq25_Tanh_Tanh_1",
    "Eq26_Tanh_Tanh_2", "Eq27_Tanh_Tanh_shape", 
     "Eq28_Tanh_Expo",
    "Eq29_Expo_Tanh_1", "Eq30_Expo_Tanh_2", "Eq31_Expo_Tanh_shape", "Eq32_Expo_Expo",
    "Eq33_Amirian_Tanh", "Eq34_Amirian_Expo"),
  data
)

Arguments

parameters

Vector -- containing the values listed below:

Pmax \(\hspace{0.1cm}\): Maximum photosynthesis rate normalised by Chl_a (sign +),

alpha: Light-saturation slope at low light level (sign +),

beta \(\hspace{0.15cm}\): Photoinhibition rate at high light level (sign +),

R \(\hspace{0.7cm}\): Dark reaction parameter (sign + OR -),

shape \(\hspace{0cm}\): Shape parameter (sign +) -- Only required for some of the models. See Details.

model_name

String -- which model? (List of available models is given in details)

data

Vector -- Containing the irradiance profile.

Value

A vector that containing the primary production (PP) calculated by a given formula provided by user.

Details

This package includes both commonly-used and recently-developed PI models. The list of models is provided below.

Table 1. Existing formulas for light-saturation curve (\(I_k = P_{\max}^B/ \alpha\)).

Name \(\hspace{0.5cm}\)	Model	Function Type	References

Eq1	\(P^B = \alpha I \hspace{1cm}\)	Linear	Linear Regression

Eq2	\(P^B = P^B_{\max}\dfrac{I+I_k - |I-I_k|}{2I_k} \hspace{1cm}\)	Bilinear	Blackman 1905

Eq3	\(P^B = P^B_{\max}\dfrac{I}{I + I_k}\)	Rectangular Hyperbola	Baly 1935

Eq4	\(P^B = P^B_{\max}\dfrac{I}{\sqrt{I^2 + I^2_k}}\)	Modified Rectangular Hyperbola	Smith 1936

Eq5	\(P^B = P^B_{\max} \ln \left( \dfrac{2I}{I_k} \right)\)	Logarithm	Talling 1957

Eq6	\(P^B = P^B_{\max}\ln \left( \dfrac{I}{I_k} + \sqrt{1 + \left(\dfrac{I}{I_k}\right)^2} \right)\)	Modified Logarithm	Vollenweider 1958

Eq7	\(P^B = P^B_{\max}(1-e^{-I/I_k})\)	Exponential	Webb et al. 1974

Eq8	\(P^B = P^B_{\max} \tanh \left( \dfrac{I}{I_k} \right)\)	Hyperbolic Tangent	Jassby et al. 1976

Eq9	\(P^B = \dfrac{P^B_{\max}}{2 \theta} \left[\mathcal{I} -\sqrt{\mathcal{I}^2 - 4 \theta(\mathcal{I} - 1)} \right]\) \(\hspace{0.5cm}\)	Non-rectangular hyperbola	Prioul et al. 1977

Eq10	\(P^B = P^B_{\max} \dfrac{I}{(I^b+I^b_k)^{1/b}}\)	Generalized Rectangular Hyperbola \(\hspace{0.5cm}\)	Bannister 1979

where \(b\) is a shape parameter, \(\theta = \dfrac{1}{1 + \exp{(-b)}}\) is a sigmoid function setting \(0< \theta <1\), and \(\mathcal{I} = \left(\dfrac{I}{I_k} + 1\right) \).

Note

The function Model_piCurve can handle all combinations of the names listed above, where the input string model_name is concerned. For instance,if any of the following forms are used, the function will set model_name = "Eq1_Blackman".

With or without spacing: Eq1Blackman, Eq1 Blackman,

Underscore or hyphen spacing: Eq1_Blackman, Eq1-Blackman,

Any form of lower/upper case: eq1_blackman, EQ1_blackman,

Missing author name: EQ1, Eq1, eQ1, eq1, 1,

Missing Eq indicator: blackman, Blackman,

Other combination of all the above: Blackmaneq1, BlackmanEq1, Blackman Eq1, eq1 Blackman, etc.

References

Will be add later

Examples

# model parameters
params <- c(Pmax = 20, alpha = 0.6, R = 0)

# generate an irradiance profile
df <- tibble::tibble(I = seq(0, 100, length = 25))

# compute the primary production profile using Baly's rectangular hyperbola model
Model_piCurve(parameters = params, model_name = "Eq3-Baly", data = df)
#>  [1]  0.000000  2.222222  4.000000  5.454545  6.666667  7.692308  8.571429
#>  [8]  9.333333 10.000000 10.588235 11.111111 11.578947 12.000000 12.380952
#> [15] 12.727273 13.043478 13.333333 13.600000 13.846154 14.074074 14.285714
#> [22] 14.482759 14.666667 14.838710 15.000000

# compute the primary production profile using Eq10 model
Model_piCurve(parameters = c(params, beta = 0.3), model_name = "Eq11_Blackman_extended_1", data = df)
#>  [1]  0.00  2.50  5.00  7.50 10.00 12.50 15.00 17.50 20.00 20.00 20.00 20.00
#> [13] 20.00 20.00 20.00 20.00 20.00 18.75 17.50 16.25 15.00 13.75 12.50 11.25
#> [25] 10.00